Fiber-Quality-Monitoring-System-main/requirements.txt
# Core Requirements
flask==2.0.3
flask-socketio==5.3.4
flask-sqlalchemy==3.0.1
python-dotenv==0.19.0
speedtest-cli==2.1.3
pandas==1.3.5
twilio==7.7.0
python-engineio==4.3.4
python-socketio==5.7.2
eventlet==0.33.3
vis.js==4.21.0
chart.js==3.5.1

# Development Extras
python-dateutil==2.8.2
ipython==8.0.1
debugpy==1.5.1

# Production
gunicorn==20.1.0